Get alocs-clip measurements/parameters closer
An interesting unintentional side-effect of this iteration was that, while the chopsticks holder was too loose to grasp the chopsticks, it was *just right* to grasp the *flatware* - more so than the ware slot itself (which, while usable, was slightly too narrow). I may revisit this idea for a future iteration.

Playing with placing the two wares immediately next to each other instead of end-to-end *also* gave me an idea for a variation on the design: placing the fork and spork front-to-back, and putting the chopsticks *outside* them, behind the spork. This would be a more compact design than the one I'm currently working toward.